Move function implementations from header to source file
